pyproject.toml
[project]
name = "mcp-remote-ssh"
version = "0.1.0"
description = "Model Context Protocol server for secure remote SSH access"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    "modelcontextprotocol>=1.0.0",
    "paramiko>=3.4.0",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",
    "asyncio-mqtt>=0.16.0",
]
requires-python = ">=3.8"
readme = "README.md"
license = {text = "MIT"}

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "docker>=7.0.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
mcp-remote-ssh = "mcp_remote_ssh.server:main"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
